Add dbu-db_init to initialize setup dir
The plan is to then implement a dbu-db_setup command that uses the setup files
to create setup sql.  Allows specification and linking to a prototype to make a
family of related setup files (the objective being that you could create a set
of related schemas, for example, each with slightly different configs).